transferred some functions to separate files, and removed some imports.
