+++ treatPrices/treatPrices.js
